refactor: extract Audio + Permissions settings panes to their own views (#355)
Continues the SettingsView pane extraction (A1 was the Diagnostics/Privacy
pane, #524). Moves the "Recording Audio" and "Permissions" GroupBoxes into
Views/Settings/SettingsAudioPanes.swift as SettingsRecordingAudioPane and
SettingsPermissionsPane, rendered in their existing body positions so the
layout is unchanged.

The pane-local helpers move with them: availableRecordingAudioSources
(recording-audio pane); microphonePermissionBlocked,
screenRecordingPermissionMissing, and permissionRecoveryRow (permissions
pane). With their only callers gone, those helpers and the now-unused
AVFoundation / CoreGraphics imports are removed from SettingsView.

Pure UI relocation: no field added/removed/rewired, every setting keeps its
SettingsStore key, no controller/validation change. Build clean (no new
warnings); full unit suite green; all 7 BugNarratorSettingsUITests pass
(one unrelated recording-controls timing flake passed on isolated re-run).

Closes #525
Refs #355
